- move context creation into the context builder
