+++ TrueDialog.Client.Standard/Client/Context/BaseContext.cs
@@ -1,43 +1,22 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Reflection;
-
-namespace TrueDialog.Context
+﻿namespace TrueDialog.Context
 {
     internal class BaseContext
     {
         protected readonly IApiCaller Api;
+        private readonly ContextBuilder m_contextBuilder;
 
         public BaseContext(IApiCaller api)
         {
             Api = api;
+            m_contextBuilder = new ContextBuilder(api);
         }
 
         protected int CurrentAccount { get { return Api.AccountId; } }
 
-        #region SubContext factory
-        protected readonly ConcurrentDictionary<string, BaseContext> m_contextDictionary = new ConcurrentDictionary<string, BaseContext>();
-        private readonly object m_mutex = new object();
-
         protected TContext GetContext<TContext>()
             where TContext : BaseContext
         {
-            lock (m_mutex)
-            {
-                string key = typeof(TContext).Name;
-                if (m_contextDictionary.ContainsKey(key))
-                    return m_contextDictionary[key] as TContext;
-                else
-                {
-                    var constructor = typeof(TContext).GetConstructor(BindingFlags.NonPublic | BindingFlags.Instance,
-                        null, new Type[] { typeof(IApiCaller) }, null);
-                    var context = (TContext)constructor.Invoke(new object[] { Api });
-
-                    m_contextDictionary[key] = context;
-                    return context;
-                }
-            }
+            return m_contextBuilder.GetContext<TContext>();
         }
-        #endregion
     }
 }
